Prevent concurrent deletes for the same persistence id
Fixes #367
