+++ core/src/main/resources/reference.conf
@@ -92,6 +92,11 @@ cassandra-journal {
   # https://docs.datastax.com/en/developer/java-driver/3.1/manual/logging/#logging-query-latencies 
   log-queries = off
 
+  # Only one outstanding delete per persistenceId is permitted. Additional deletes are queued.
+  # After max-concurrent-deletes is hit for a single persistenceId deletes are failed automatically.
+  # Deleting should be used with snapshots for efficiency for recovery thus deleting should be infrequent
+  max-concurrent-deletes = 10
+
   # Cassandra driver connection pool settings
   # Documented at http://docs.datastax.com/en/developer/java-driver/latest/manual/pooling/
   connection-pool {
